# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "ISP management system.",
        "title": "ISP Management System API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1/",
    "paths": {
        "/api/v1/auth/reset_password": {
            "post": {
                "description": "returns the ID, Email of the currently loggedin user.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "/api/v1/auth/reset_password ResetPassword allows you to reset your password.",
                "parameters": [
                    {
                        "description": "User Data",
                        "name": "user",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/controllers.ResetPasswordInput"
                        }
                    }
                ]
            }
        },
        "/api/v1/auth/whoami": {
            "get": {
                "description": "returns the ID, Email of the currently loggedin user.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "/api/v1/auth/whoami returns the basic details (id, email) of the logged user."
            }
        },
        "/api/v1/signup": {
            "post": {
                "description": "Signup/Register/Add users",
                "summary": "Signup/Register/Add Users"
            }
        }
    },
    "definitions": {
        "controllers.ResetPasswordInput": {
            "type": "object",
            "required": [
                "confirmNewPassword",
                "currentPassword",
                "newPassword"
            ],
            "properties": {
                "confirmNewPassword": {
                    "type": "string"
                },
                "currentPassword": {
                    "type": "string"
                },
                "newPassword": {
                    "type": "string"
                }
            }
        }
    }
}